refactor: spaceFooter 분리
