+++ frontend/src/components/specific/space/spaceHeader/SpaceHeader.styles.ts
