+++ frontend/src/components/specific/space/spaceHeader/managerSpaceHeader/ManagerHeader.styles.ts
